+++ renderers/__init__.py
@@ -1,6 +1,7 @@
 from .catalog import render_catalog_grid, render_pigsty
 from .common import draw_bold_text, fit_card_image, get_text_size
 from .pig_card import PigCardLayout, render_pig_card
+from .roast import render_roast_card
 from .weekly import WeeklyEntry, render_weekly_summary
 
 __all__ = [
@@ -12,5 +13,6 @@ __all__ = [
     "render_catalog_grid",
     "render_pig_card",
     "render_pigsty",
+    "render_roast_card",
     "render_weekly_summary",
 ]